Refactor: 링크 저장 기능 책임 분리
