+++ src/main/java/cmc/blink/domain/link/business/LinkInfoExtractor.java
@@ -1,0 +1,7 @@
+package cmc.blink.domain.link.business;
+
+import cmc.blink.domain.link.presentation.dto.LinkResponse;
+
+public interface LinkInfoExtractor {
+    LinkResponse.LinkInfo extractInfo(String domain, String url) throws Exception;
+}
