--- src/main/java/cmc/blink/domain/link/business/LinkScrapper.java
@@ -1,10 +1,0 @@
-package cmc.blink.domain.link.business;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class LinkScrapper {
-
-}
